Support zero trigger levels
By adding float support to AbstractCommand::parseNumber() and updating
DsoCommand to use it.

Also added support for, and standardised on, the Unicode lowercase mu
(U+03BC), in addition to the already-supported Unicode compatibility
micro symbol (U+00B5). Users may use either, but Dokit will (continue
to) always use the lowercase mu (as recommended by Unicode).

Note, the micro and mu support does not work on Windows with Qt5. Can
still use the ASCII 'u' character, as most users would. This is not a
new limitation, just one that has been revealed by recently added unit
tests.

Relates-to: https://github.com/pcolby/dokit/issues/46

Squashed commit of the following:

commit afe10af0eff3a8c7e850417f2734cb26383b5804
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:56:49 2026 +1100

    Restore Linux and macOS builds

commit cde4a8ba80db2f07ba43d126b66578746c07905d
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:43:16 2026 +1100

    Don't expect failures with MinGW

    Only MSVC has issues.

commit ca395a1a727ed961090c01ba973d5af94c33368e
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:33:44 2026 +1100

    Restore Linux and macOS builds

    And update translation files.

commit 51f39e58f61d46980325a4af2cb5b42c69452131
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:32:05 2026 +1100

    Remove temporary debug logging

commit da8563610b0bf0b1a50ccdb8fad57bdfcd018a69
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:24:13 2026 +1100

    Don't truncate to integer prematurely

commit ea3b9f0b9539c2fadd3c86ad7f1de6661c3d967b
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:13:59 2026 +1100

    Make everything explicitly double

    To see if it helps with the MinGW rounding issue.

commit 5c5bc3377378b956ede00cd062fe4d1f08a17936
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 21:00:54 2026 +1100

    Remove potential double-to-integer truncation

    We round later anyway, if necessary.

commit 1883ea1e7f112afd258f5e0c05be7691177382cb
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 20:48:28 2026 +1100

    Add more Windows debugging

commit d02d607358f6abeccdd1f827944d7e92bbae56f8
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 20:20:06 2026 +1100

    Correct data labels

commit 4dad093ab3d8ec67140b15801cc429f562706f6a
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 20:01:26 2026 +1100

    Add exception for the known buggy versions

commit 6516e986833037f6a9ea861964619ca1e771fb88
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 19:33:48 2026 +1100

    Add more Windows debugging

commit 6b11abe4eceb2c332221dcf58b8f00e9e2181da3
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 19:10:38 2026 +1100

    Diagnose a Windows-specific issue

commit 7652ea0ef644505441e1458ed396c227cfa2cf40
Author: Paul Colby <git@colby.id.au>
Date:   Wed Apr 1 17:08:45 2026 +1100

    Support Qt's earlier than 5.10.

commit acf2a15efce8a9f9ec85aea1b32753bb73481d02
Author: Paul Colby <git@colby.id.au>
Date:   Tue Mar 31 19:18:30 2026 +1100

    Support zero trigger levels

commit cf178745b24a6bf497ab658299f557b2bd656fb0
Author: Paul Colby <git@colby.id.au>
Date:   Tue Mar 31 19:18:22 2026 +1100

    Add float support to AbstractCommand::parseNumber()

    And update DsoCommand to use it.

    Also added support for, and standardised on, the Unicode lowercase
    mu (U+03BC), in addition to the already-supported Unicode
    compatibility micro symbol (U+00B5). Users may use either, but
    Dokit will (continue to) always use the lowercase mu (as
    recommended by Unicode).

+++ src/cli/metercommand.cpp
@@ -119,7 +119,7 @@ QStringList MeterCommand::processOptions(const QCommandLineParser &parser)
                 rangeOptionValue = parseNumber<std::ratio<1>>(value, u"ohms"_s);
                 break;
             case MultimeterService::Mode::Capacitance:
-                rangeOptionValue = parseNumber<std::nano>(value, u"F"_s, 500); // pF.
+                rangeOptionValue = parseNumber<std::nano>(value, u"F"_s, (quint32)500); // pF.
                 break;
             default:
                 qCInfo(lc).noquote() << tr("Ignoring range value: %1").arg(value);
